Add files via upload
The Frontend Image Carousel project is a web-based application that allows users to view a collection of images in a visually appealing and interactive manner. The images are displayed in a slideshow format, with users able to navigate between images using arrow buttons or by clicking on thumbnail previews of the images.